refactr node fetching target with fallback

+++ src/trigger/node/action/trigger-action-remove-item.ts
@@ -3,10 +3,12 @@ import { TriggerNode } from "../trigger-node";
 
 class RemoveItemTriggerAction extends TriggerNode<typeof removeItemSchema> {
     async execute(): Promise<void> {
-        const target = (await this.get("target")) ?? this.target;
-        const item = this.getExistingItem(target, await this.get("item"));
+        const target = await this.getTarget("target");
 
-        await item?.delete();
+        if (target) {
+            const item = this.getExistingItem(target, await this.get("item"));
+            await item?.delete();
+        }
 
         return this.send("out");
     }
